Vastly simplified the decision tree.
Originally, I took a lot of size measurements to try to decide
if I should use PNG or JPEG. But it turns out you should pretty
much always use JPEG unless the webp is lossless or uses features
that aren't in JPEG like alpha channels or animation.
So the code does only that.